+++ lib/Braintree/PaymentInstrumentType.php
@@ -11,5 +11,6 @@ class PaymentInstrumentType
     const ANDROID_PAY_CARD    = 'android_pay_card';
     const VENMO_ACCOUNT       = 'venmo_account';
     const US_BANK_ACCOUNT     = 'us_bank_account';
+    const IDEAL_PAYMENT       = 'ideal_payment';
 }
 class_alias('Braintree\PaymentInstrumentType', 'Braintree_PaymentInstrumentType');