+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add `IDEAL_PAYMENT` to `PaymentInstrumentType`
+
 ## 3.21.2
 * Add option to disable Accept-Encoding: gzip header for Google App Engine
 * Fix a bug where `merchantAccount->all` would attempt to fetch too many pages of merchant accounts